requirements.txt
## `requirements.txt`

# core
llama-cpp-python>=0.1.50          # or replace with your preferred local LLM binding
PyYAML>=6.0
pydantic>=1.10

# tools
requests>=2.28
beautifulsoup4>=4.12
pyppdf>=0.0.7
PyPDF2>=3.0
gspread>=5.8.0
oauth2client>=4.1.3

# small utilities
streamlit>=1.25
flask>=2.2
pandas>=2.0
schedule>=1.1.0

# optional (vector db, embeddings)
faiss-cpu>=1.7.4; python_version >= "3.8"
sentence-transformers>=2.2.2
